tkinter_CRUD/requirements.txt

ponyORM